requirements.txt
Flask==3.1.0
flask_wtf==1.2.2
python-dotenv==1.0.1
Requests==2.32.3
